app/requirements.txt
aiohttp==3.4.4
appdirs==1.4.3
async-timeout==3.0.1
attrs==18.2.0
azure-storage
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
multidict==4.4.2
PyJWT==1.6.4
PySocks==1.6.8
pytz==2018.7
requests==2.20.1
six==1.11.0
spotify==0.1.9
toml==0.10.0
twilio==6.19.2
urllib3==1.24.1
Werkzeug==0.14.1
yarl==1.2.6
